Plot: reverse to march other monitoring programs behavior
Change the plot order to mimic the other monitoring programs behavior.
The consensus seems to be that the most recent point is displayed on the
right of the plot and oldest on the left.

Closes #69 as this is now the default.
The command line option "-r, --reverse-abs" is provided to reverse the
plot, getting the pre-commit behaviour if preferred.