Fix: filter out nvidia utilization samples
The function nvmlDeviceGetProcessUtilization might return a number of
samples that exceeds the running number of processes on the GPU.
Furthermore, most of the returned samples are filled with values that,
either do not make sense (e.g. >100% utilization rate) or with a
timestamp in the past of what we asked for.

Fixes #110.