extract_gpuinfo_amdgpu: Support per-process GPU usage
This is via the fdinfo interface. The enumeration of all the visible
fds on a system is slightly expensive, but I'm not sure how to avoid it.

The logic is partly adapted from unmerged intel_gpu_top code [1],
adapted for AMDGPU. i915 kernel changes has not been merged into
mainline yet, AFAICT.

[1] https://patchwork.freedesktop.org/series/100571/