cmake: use check_linker_flag for checking linker flags
check_c_compiler_flag does not actually pass the flag to the linker, and
so the test file successfully compiles for any linker flag. This
caused the -Wl,z,relro flag to be passed to the compiler even if it was
not supported.

check_c_compiler_flag was introduced in cmake 3.18, so this commit also
bumps the minimum required cmake version to 3.18 (released in 2020).