apple: dynamic memory info
There are two main cases to consider here: When the GPU has unified memory
(shared with the host) and when not. In the former, we can query the amount
of system memory that is allocated by the GPU to get an indicative value of
what's being used. For the total amount of memory, we can just get the system
memory.

It is not really possible to get these values when the GPU does not have
unified memory. In this case we can only get the recommended maximum working
set size (this is what MoltenVK returns in this case). TODO is to find out if
we can find out how much memory is allocated on such devices.