apple: implement fetching running processes
It happens that we can get this information from the IO Registry. The
service that is associated with an MTLDevice's registryID has a bunch
of children of class AGXDeviceUserClient. These contain some basic
information about the process, such as the PID and GPU time used.

Unfortunately there does not seem to be a dedicated metric for the amount
of memory held by the process, nor does there seem to be a way to
tell whether a process is using compute or graphics.

This commit only fetched the PID of running processes. More advanced
information is TODO.