apple: implement get_process_info
All the required information can be fetched through proc_taskinfo.
The time scale of pti_total_user and pti_total_system must be multiplied
with a scalar obtained from the system to translate the value into
nanoseconds. According to the source of htop, there is an error on M1
when running under Rosetta, but that is not (yet) implemented. Also, the
source mentions HAVE_MACH_TIMEBASE_INFO, implying that this call is not
always available. This test is also not yet written.