apple: implement get_command_from_pid
KERN_PROCARGS2 can be used to query the initial
stack that is passed to a program. We can fetch the argv manually
from there. Supposedly this is how ps implements this also.

One caveat is that this only works for processes for which we
have sufficient privilege to query the procargs. The only reason that
ps is able to fetch it is because its running with setuid. If this
information is really desired nvtop can be executed with root, but it
is not really critical.