Toggle Panfrost's sysfs profiling hook after DRM handle is retrieved
Otherwise nvtop will attempt to toggle them even on systems for which
nvtop was built with Panfrost support, but which possess no Mali GPUs.

This also makes sure the right sysfs profiling hook is touched for every
single Mali GPU in the system, even on SoCs with more than one of
these (yes, those do exist).

Signed-off-by: Adrián Larumbe <adrian.larumbe@collabora.com>